requirements.txt
accelerate==1.1.1
bitsandbytes==0.44.1
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
cloudpickle==3.1.0
cmake==3.31.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
derivative==0.6.3
distlib==0.3.9
Farama-Notifications==0.0.4
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
gymnasium==1.0.0
-e ./gymnasium_envs
huggingface-hub==0.26.2
identify==2.6.3
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.2
nodeenv==1.9.1
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
pre_commit==4.0.1
psutil==6.1.0
pygame==2.6.1
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyserial==3.5
pysindy==1.7.5
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
scs==3.2.7
setuptools==75.1.0
six==1.16.0
stable_baselines3==2.4.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1+cu124 @ https://download.pytorch.org/whl/cu124
torchaudio==2.5.1+cu124 @ https://download.pytorch.org/whl/cu124
torchvision==0.20.1+cu124 @ https://download.pytorch.org/whl/cu124
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.28.0
wheel==0.44.0
zipp==3.21.0
